+++ src/Funds/CoinExchange.Funds.Domain.Model/DepositAggregate/DepositLimit.cs
@@ -21,7 +21,7 @@ namespace CoinExchange.Funds.Domain.Model.DepositAggregate
         /// <summary>
         /// Initializes a new instance of the <see cref="T:System.Object"/> class.
         /// </summary>
-        public DepositLimit(string tierLevel, double dailyLimit, double monthlyLimit)
+        public DepositLimit(string tierLevel, decimal dailyLimit, decimal monthlyLimit)
         {
             TierLevel = tierLevel;
             DailyLimit = dailyLimit;
@@ -41,11 +41,11 @@ namespace CoinExchange.Funds.Domain.Model.DepositAggregate
         /// <summary>
         /// Monthly Limit
         /// </summary>
-        public double MonthlyLimit { get; private set; }
+        public decimal MonthlyLimit { get; private set; }
 
         /// <summary>
         /// Daily Limit
         /// </summary>
-        public double DailyLimit { get; private set; }
+        public decimal DailyLimit { get; private set; }
     }
 }
