+++ src/Funds/CoinExchange.Funds.Domain.Model/DepositAggregate/IDepositLimitEvaluationService.cs
@@ -12,32 +12,32 @@ namespace CoinExchange.Funds.Domain.Model.DepositAggregate
         /// Evaluate the limit for deposit and signify if current transaction is within the deposit limits
         /// </summary>
         /// <returns></returns>
-        bool EvaluateDepositLimit(double amountInUsd, IList<Ledger> depositLedgers, DepositLimit depositLimit,
-            double bestBidPrice, double bestAskPrice);
+        bool EvaluateDepositLimit(decimal amountInUsd, IList<Ledger> depositLedgers, DepositLimit depositLimit,
+            decimal bestBidPrice, decimal bestAskPrice);
 
         /// <summary>
         /// Daily Deposit Limit
         /// </summary>
-        double DailyLimit { get; }
+        decimal DailyLimit { get; }
 
         /// <summary>
         /// Daily Deposit Limit that has been used in the last 24 hours
         /// </summary>
-        double DailyLimitUsed { get; }
+        decimal DailyLimitUsed { get; }
 
         /// <summary>
         /// Monthly Deposit Limit
         /// </summary>
-        double MonthlyLimit { get; }
+        decimal MonthlyLimit { get; }
 
         /// <summary>
         /// Monthly Deposit Limit that has been used in the last 30 days
         /// </summary>
-        double MonthlyLimitUsed { get; }
+        decimal MonthlyLimitUsed { get; }
 
         /// <summary>
         /// Maximum Deposit amount allowed to the user at this moment.. Deposit should be kept 5-10% lower than this limit
         /// </summary>
-        double MaximumDeposit { get; }
+        decimal MaximumDeposit { get; }
     }
 }
